feat(account-lib): implement validityWindow and sequenceId for Sol
implement validityWindow and sequenceId for Sol

stlx-9890
